Separate useUser from common.test
